eslint, prettier, solhint

+++ contracts/interface/IOption.sol
@@ -2,10 +2,9 @@
 
 pragma solidity ^0.6.0;
 
-import { Bitcoin } from "../types/Bitcoin.sol";
+import {Bitcoin} from '../types/Bitcoin.sol';
 
 interface IOption {
-
     function initialize(
         uint8 _decimals,
         uint256 _expiryTime,
@@ -18,9 +17,15 @@ interface IOption {
 
     function obligation() external returns (address);
 
-    function mint(address from, address to, uint256 amount, bytes20 btcHash, Bitcoin.Script format) external;
+    function mint(
+        address from,
+        address to,
+        uint256 amount,
+        bytes20 btcHash,
+        Bitcoin.Script format
+    ) external;
 
-    function requestExercise(address seller, uint satoshis) external;
+    function requestExercise(address seller, uint256 satoshis) external;
 
     function executeExercise(
         address seller,
@@ -31,5 +36,5 @@ interface IOption {
         bytes calldata rawtx
     ) external;
 
-    function refund(uint amount) external;
+    function refund(uint256 amount) external;
 }
